refactor: WinningType 을 view에서 domain으로 이동
