+++ src/test/java/blackjack/domain/WinningTypeDiscriminatorTest.java
@@ -3,7 +3,7 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.gambler.Name;
-import blackjack.view.WinningType;
+
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
